REST-api/requirements.txt
django==1.10.2
djangorestframework
requests